App completed

+++ app/src/main/java/com/example/marvel/MarvelAPI.java
@@ -1,19 +1,17 @@
 package com.example.marvel;
 
-import java.util.List;
+import com.example.marvel.Models.Wrapper;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
-import retrofit2.http.Headers;
-import retrofit2.http.Path;
 import retrofit2.http.Query;
 
 public interface MarvelAPI {
 
 
     @GET("v1/public/characters")
-    Call<Wrapper> getCharacters(@Query("ts") int ts,
-                             @Query("hash") String hash,
-                             @Query("nameStartsWith") String character,
-                             @Query("apikey") String apikey);
+    Call<Wrapper> getCharacters(@Query("ts") long ts,
+                                @Query("hash") String hash,
+                                @Query("nameStartsWith") String character,
+                                @Query("apikey") String apikey);
 }
